fix: Remove indirect import of AWS package
