+++ pkg/common/common.go
@@ -2,14 +2,11 @@ package common
 
 import (
 	"fmt"
-	"net/http"
-	"net/url"
 	"os"
 	"regexp"
 	"sort"
 	"strings"
 
-	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/vmware-tanzu/velero/pkg/types"
 	corev1 "k8s.io/api/core/v1"
 )
@@ -208,23 +205,6 @@ func CCOWorkflow() bool {
 		return true
 	}
 	return false
-}
-
-// StripDefaultPorts removes port 80 from HTTP URLs and 443 from HTTPS URLs.
-// Defer to the actual AWS SDK implementation to match its behavior exactly.
-func StripDefaultPorts(fromUrl string) (string, error) {
-	u, err := url.Parse(fromUrl)
-	if err != nil {
-		return "", err
-	}
-	r := http.Request{
-		URL: u,
-	}
-	request.SanitizeHostForHeader(&r)
-	if r.Host != "" {
-		r.URL.Host = r.Host
-	}
-	return r.URL.String(), nil
 }
 
 // GetImagePullPolicy get imagePullPolicy for a container, based on its image, if an override is not provided.
